chore(CheckInTab): Bring the mutation and state definition one level up to the drawer component
