+++ web/src/components/CheckInTab/CheckInTab.tsx
@@ -1,54 +1,27 @@
-import { useState } from 'react'
-
-import { FindWorkRequestQuery } from 'types/graphql'
-
-import { useMutation } from '@redwoodjs/web'
-import { toast } from '@redwoodjs/web/toast'
+import { Dispatch, SetStateAction } from 'react'
 
 import ButtonWithLoader from 'src/components/ButtonWithLoader'
 import TempAgencyWorker from 'src/components/TempAgencyWorker'
 import TimePicker from 'src/components/TimePicker'
 import { Badge } from 'src/components/ui/badge'
 import { Button } from 'src/components/ui/button'
-import { QUERY } from 'src/components/WorkRequestCell'
-
-const CHECK_IN_GQL = gql`
-  mutation CheckIn($id: String!, $input: UpdateShiftInput!) {
-    updateShift(id: $id, input: $input) {
-      id
-    }
-  }
-`
 
 type CheckInTabProps = {
-  shift: FindWorkRequestQuery['workRequest']['shifts'][0]
+  checkInAt: Date
+  setCheckInAt: Dispatch<SetStateAction<Date>>
+  loading: boolean
+  handleCheckIn: () => void
 }
 
-const CheckInTab = ({ shift }: CheckInTabProps) => {
-  const [checkedInAt, setCheckedInAt] = useState(
-    shift.checkedInAt ? new Date(shift.checkedInAt) : new Date()
-  )
-  const [checkIn, { loading }] = useMutation(CHECK_IN_GQL, {
-    onCompleted: () => toast.success('Ingecheckt met succes'),
-    onError: (e) => toast.error(e.message),
-    refetchQueries: [QUERY],
-  })
-
+const CheckInTab = ({
+  checkInAt,
+  setCheckInAt,
+  loading,
+  handleCheckIn,
+}: CheckInTabProps) => {
   function handleClickNow() {
-    setCheckedInAt(new Date())
+    setCheckInAt(new Date())
   }
-
-  async function handleConfirm() {
-    await checkIn({
-      variables: {
-        id: shift.id,
-        input: {
-          checkedInAt,
-        },
-      },
-    })
-  }
-
   return (
     <>
       <div className="my-4 flex h-[250px] flex-col justify-between">
@@ -67,11 +40,11 @@ const CheckInTab = ({ shift }: CheckInTabProps) => {
               Nu
             </Button>
           </h3>
-          <TimePicker date={checkedInAt} onDateChange={setCheckedInAt} />
+          <TimePicker date={checkInAt} onDateChange={setCheckInAt} />
         </div>
         <ButtonWithLoader
           loading={loading}
-          onClick={handleConfirm}
+          onClick={handleCheckIn}
           className="bg-accent/80 text-black hover:bg-accent hover:text-black sm:mx-auto"
         >
           Inchecken
